+++ NEWS.md
@@ -1,5 +1,6 @@
 # ppseq (development version)
 
+* adds informative messaging when no results returned from `optimize_design()` (Closed Issue #14)
 * upgraded package website to bootstrap 5 and changed some styling
 * added two-sample randomized trial vignette
 * added `calc_next()` function to calculate probability of response in next patient (Closed Issue #2)